// frontend/src/app/learncoreservices/models/student-detail-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StudentDetailResponse {
  accountLocked?: boolean;
  createdDate?: string;
  email?: string;
  enabled?: boolean;
  firstname?: string;
  lastModifiedDate?: string;
  lastname?: string;
  level?: number;
  profilePictureUrl?: string;
  referCode?: string;
  streakDays?: number;
  subscriptionType?: string;
  tenantUuid?: string;
  username?: string;
  uuid?: string;
  xp?: number;
}